# api/openapi.yaml
...
openapi: "3.0.2"
info:
  title: "Configuration Framework Service"
  version: "0000.0000.0000"
  description: >
    The Configuration Framework Service (CFS) manages the launch of Ansible Execution Environments
    for image customization, node personalization, and node reconfiguration. CFS manages
    the Ansible staging container that pulls Ansible play/role content and inventory (optional)
    from the git server and launches the Ansible Execution Environment.


    CFS includes the following components:

      * CFS REST API
      * A Kubernetes operator running on the management services infrastructure to handle
      the lifecycle and reporting of configuration sessions.
      * Pre-packaged Ansible Execution Environment(s) (AEE) with values tuned for performant
      configuration for executing Ansible playbooks against Cray compute and user access nodes.


    CFS uses a Git version control server running in the management services infrastructure
    for management of the configuration manifest lifecycle.


      The CFS API allows an administrator to customize the compute and user access nodes
      in the following ways:
    * Customize the bootable images prior to their use on the system. This process is called
      image customization. CFS uses IMS to stage images in an ssh container and then modifies
      one or more images using Ansible.

    * Customize live nodes during boot or post-boot. This process is called node personalization.
...

    ## Resources

...


...


...


...

...
servers:
...
components:
  requestBodies:
...
    V3SourceCreateRequest:
      description: A source
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V3SourceCreateData'
    V3SourceRestoreRequest:
      description: A source
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V3SourceRestoreData'
    V3SourceUpdateRequest:
      description: A source
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V3SourceUpdateData'
  responses:
...
    V3ConfigurationDataCollection:
      description: A collection of configurations
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V3ConfigurationDataCollection'
    V3SourceData:
      description: A single source
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V3SourceData'
    V3SourceDataCollection:
      description: A collection of sources
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/V3SourceDataCollection'
    ResourceDeleted:
      description: The resource was deleted.
    # Errors
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ResourceNotFound:
      description: The resource was not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    ConflictingSourceName:
      description: A source with the same name already exists
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    V3NextData:
      description: Information for requesting the next page of data
      type: object
      properties:
        limit:
          type: integer
        after_id:
          type: string
      nullable: true
      additionalProperties: true  # This can contain the parameters of any endpoint this is returned from
    ## OPTIONS ENDPOINT ##
...
    V3AdditionalInventoryLayer:
      description: |
        An inventory reference to include in a set of configurations.
        Either clone_url or source must be specified -- it is required to specify one,
        but they are mutually exclusive.
      type: object
      properties:
        name:
          type: string
          description: The name of the inventory layer.
          example: sample-inventory
          minLength: 1
          maxLength: 45
        clone_url:
          type: string
          description: The clone URL of the configuration content repository.
          example: https://vcs.domain/vcs/org/inventory.git
          pattern: '^[^\s;]*$'
        source:
          type: string
          description: A CFS source with directions to the configuration content repository
        commit:
          type: string
          description: The commit hash of the configuration repository when the state is set.
          pattern: '^[a-f0-9]{40}$'
        branch:
          type: string
          description: |
            The repository branch to use. This will automatically set `commit` to master on the branch
            when the configuration is added.
          pattern: '^[^\s;]*$'
      additionalProperties: false
...
    V3ConfigurationLayer:
      description: |
        A single desired configuration state for a component.
        Either clone_url or source must be specified -- it is required to specify one,
        but they are mutually exclusive.
      type: object
      properties:
        name:
          type: string
          description: The name of the configuration layer.
          example: sample-config
          minLength: 1
          maxLength: 45
        clone_url:
          type: string
          description: The clone URL of the configuration content repository.
          example: https://api-gw-service-nmn.local/vcs/cray/config-management.git
          pattern: '^[^\s;]*$'
        source:
          type: string
          description: A CFS source with directions to the configuration content repository
        playbook:
          type: string
          description: The Ansible playbook to run.
          example: site.yml
          pattern: '^[^\s;]*$'
        commit:
          type: string
          description: The commit hash of the configuration repository when the state is set.
          pattern: '^[a-f0-9]{40}$'
        branch:
          type: string
          description: |
            The configuration branch to use.  This will automatically set commit to master on the branch
            when the configuration is added.
          pattern: '^[^\s;]*$'
        special_parameters:
          type: object
          description: |
            Optional parameters that do not affect the configuration content or are only used in
            special circumstances.
          properties:
            ims_require_dkms:
              type: boolean
              description: |
                If true, any image customization sessions that use this configuration will enable DKMS in IMS.
          additionalProperties: false
      required: [playbook]
      additionalProperties: false
...
    V3ConfigurationData:
      description: A collection of ConfigurationLayers.
      type: object
      properties:
        name:
          type: string
          description: The name of the configuration.
          example: sample-config
          readOnly: true
        description:
          type: string
          description: A user-defined description. This field is not used by CFS.
        tenant_name:
          type: string
          description: The name of a tenant that owns this set of configuration.
          example: blue-tenant
          required: false
        last_updated:
          type: string
          description: The date/time when the state was last updated in RFC 3339 format.
          example: '2019-07-28T03:26:00Z'
          format: date-time
          readOnly: true
        layers:
          type: array
          description: A list of ConfigurationLayer(s).
          items:
            $ref: '#/components/schemas/V3ConfigurationLayer'
        additional_inventory:
          $ref: '#/components/schemas/V3AdditionalInventoryLayer'
      additionalProperties: false
...
    V3ConfigurationDataCollection:
      description: A collection of configuration data.
      type: object
      properties:
        configurations:
          type: array
          items:
            $ref: '#/components/schemas/V3ConfigurationData'
        next:
          $ref: '#/components/schemas/V3NextData'
    ## COMPONENTS ENDPOINT ##
...
    V3SourceCredentials:
      description: Information on a secret containing the username and password for accessing the git content
      type: object
      properties:
        authentication_method:
          type: string
          description: The git authentication method used.
          enum:
            - password
#            - token
#            - key
        secret_name:
          type: string
          description: The name of the credentials vault secret.
          readOnly: true
        username:
          type: string
          description: The username for authenticating to git
          writeOnly: true
        password:
          type: string
          description: The password for authenticating to git
          writeOnly: true
      additionalProperties: false
    V3SourceCert:
      description: Information on a configmap containing a CA certificate for authenticating to git
      type: object
      properties:
        configmap_name:
          type: string
          description: The name of the configmap containing a necessary CA cert.
          minLength: 1
        configmap_namespace:
          type: string
          description: The namespace of the CA cert configmap in kubernetes.
      nullable: true
      additionalProperties: false
      required:
        - configmap_name
    V3SourceData:
      description: Information for retrieving git content from a source.
      type: object
      properties:
        name:
          type: string
          description: The name of the source.  This field is optional and will default to the clone_url if not specified.
          example: sample-source
        description:
          type: string
          description: A user-defined description. This field is not used by CFS.
        last_updated:
          type: string
          description: The date/time when the state was last updated in RFC 3339 format.
          example: '2019-07-28T03:26:00Z'
          format: date-time
          readOnly: true
        clone_url:
          type: string
          description: The url to access the git content
        credentials:
          $ref: '#/components/schemas/V3SourceCredentials'
        ca_cert:
          $ref: '#/components/schemas/V3SourceCert'
      additionalProperties: false
    V3SourceDataCollection:
      description: A collection of source data.
      type: object
      properties:
        sources:
          type: array
          items:
            $ref: '#/components/schemas/V3SourceData'
        next:
          $ref: '#/components/schemas/V3NextData'
    V3SourceCreateCredentials:
      description: Information for retrieving the git credentials
      type: object
      properties:
        authentication_method:
          type: string
          description: The git authentication method used.
          enum:
            - password
          default: password
        username:
          type: string
          description: The username for authenticating to git
          minLength: 1
          writeOnly: true
        password:
          type: string
          description: The password for authenticating to git
          minLength: 1
          writeOnly: true
      additionalProperties: false
      required:
        - username
        - password
    V3SourceRestoreCredentials:
      description: Information on a secret containing the username and password for accessing the git content
      type: object
      properties:
        authentication_method:
          type: string
          description: The git authentication method used.
          enum:
            - password
        secret_name:
          type: string
          description: The name of the credentials vault secret.
          minLength: 1
      required:
        - authentication_method
        - secret_name
      additionalProperties: false
    V3SourceCreateData:
      description: Information for retrieving git content from a source.
      type: object
      properties:
        name:
          type: string
          description: The name of the source.  This field is optional and will default to the clone_url if not specified.
          example: sample-source
          minLength: 1
        description:
          type: string
          description: A user-defined description. This field is not used by CFS.
        clone_url:
          type: string
          description: The url to access the git content
          minLength: 1
        credentials:
          $ref: '#/components/schemas/V3SourceCreateCredentials'
        ca_cert:
          $ref: '#/components/schemas/V3SourceCert'
      additionalProperties: false
      required:
        - clone_url
        - credentials
    V3SourceRestoreData:
      description: Information for retrieving git content from a source.
      type: object
      properties:
        description:
          type: string
          description: A user-defined description. This field is not used by CFS.
        clone_url:
          type: string
          description: The url to access the git content
          minLength: 1
        credentials:
          $ref: '#/components/schemas/V3SourceRestoreCredentials'
        ca_cert:
          $ref: '#/components/schemas/V3SourceCert'
      additionalProperties: false
      required:
        - clone_url
        - credentials
    V3SourceUpdateData:
      description: Information for retrieving git content from a source.
      type: object
      properties:
        description:
          type: string
          description: A user-defined description. This field is not used by CFS.
        clone_url:
          type: string
          description: The url to access the git content
          minLength: 1
        credentials:
          $ref: '#/components/schemas/V3SourceCredentials'
        ca_cert:
          $ref: '#/components/schemas/V3SourceCert'
      additionalProperties: false
    ## OTHER ##
    ProblemDetails:
      description: An error response for RFC 7807 problem details.
      type: object
      properties:
        type:
          description:
            Relative URI reference to the type of problem which includes human
            readable documentation.
          type: string
          format: uri
          default: "about:blank"
        title:
          description:
            Short, human-readable summary of the problem, should not change by
            occurrence.
          type: string
        status:
          description: HTTP status code
          type: integer
          example: 400
        instance:
          description: A relative URI reference that identifies the specific
            occurrence of the problem
          format: uri
          type: string
        detail:
          description:
            A human-readable explanation specific to this occurrence of the
            problem. Focus on helping correct the problem, rather than giving
            debugging information.
          type: string
paths:
...
  /v3/configurations:
    get:
      summary: Retrieve a collection of configurations
      tags:
        - configurations
        - v3
      x-openapi-router-controller: cray.cfs.api.controllers.configurations
      description: >-
        Retrieve the full collection of configurations in the form of a
        ConfigurationArray.
      operationId: get_configurations_v3
      parameters:
        - name: limit
          schema:
            type: integer
            minimum: 1
          in: query
          description: >-
            When set, CFS will only return a number of configurations up to this limit.  Combined with after_id, this
            enables paging across results
        - name: after_id
          schema:
            type: string
          in: query
          description: >-
            When set, CFS will only return the configurations after the configuration specified.  Combined with limit, this
            enables paging across results.
        - name: in_use
          schema:
            type: boolean
          in: query
          description: >-
            Query for only configurations that are currently referenced by components.
      responses:
        200:
          $ref: '#/components/responses/V3ConfigurationDataCollection'
        400:
          $ref: '#/components/responses/BadRequest'
...
  /v3/sources:
    get:
      summary: Retrieve a collection of sources
      tags:
        - sources
        - v3
      x-openapi-router-controller: cray.cfs.api.controllers.sources
      description: >-
        Retrieve the full collection of sources in the form of a
        SourceArray.
      operationId: get_sources_v3
      parameters:
        - name: limit
          schema:
            type: integer
            minimum: 1
          in: query
          description: >-
            When set, CFS will only return a number of sources up to this limit.  Combined with after_id, this
            enables paging across results
        - name: after_id
          schema:
            type: string
          in: query
          description: >-
            When set, CFS will only return the configurations after the source specified.  Combined with limit, this
            enables paging across results.
        - name: in_use
          schema:
            type: boolean
          in: query
          description: >-
            Query for only sources that are currently referenced by configurations.
      responses:
        200:
          $ref: '#/components/responses/V3SourceDataCollection'
        400:
          $ref: '#/components/responses/BadRequest'
    post:
      summary: Add a single source
      tags:
        - sources
        - v3
      x-openapi-router-controller: cray.cfs.api.controllers.sources
      description: Add a source to CFS
      operationId: post_source_v3
      requestBody:
        $ref: '#/components/requestBodies/V3SourceCreateRequest'
      responses:
        201:
          $ref: '#/components/responses/V3SourceData'
        400:
          $ref: '#/components/responses/BadRequest'
        409:
          $ref: '#/components/responses/ConflictingSourceName'
  /v3/sources/{source_id}:
    get:
      summary: Retrieve a single source
      tags:
        - sources
        - v3
      x-openapi-router-controller: cray.cfs.api.controllers.sources
      description: Retrieve the given source
      operationId: get_source_v3
      responses:
        200:
          $ref: '#/components/responses/V3SourceData'
        404:
          $ref: '#/components/responses/ResourceNotFound'
    patch:
      summary: Update the commits for a source
      tags:
        - sources
        - v3
      x-openapi-router-controller: cray.cfs.api.controllers.sources
      description: Updates a CFS source
      operationId: patch_source_v3
      requestBody:
        $ref: '#/components/requestBodies/V3SourceUpdateRequest'
      responses:
        200:
          $ref: '#/components/responses/V3SourceData'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/ResourceNotFound'
    post:
      summary: Restore a source
      tags:
        - sources
        - v3
        - cli_ignore
      x-openapi-router-controller: cray.cfs.api.controllers.sources
      description:  >-
        Restore a CFS source by providing the name of the Vault secret that contains the credentials.
        This does NOT create the secret in Vault, nor does it validate that the secret exists.
        This is intended to be used to restore CFS data in the case that it is lost or corrupted.
        This will NOT recover a source that has been deleted using CFS, because when CFS deletes a
        source, it also deletes its corresponding Vault secret.
        NOTE: This action is not available prior to CFS 1.23
      operationId: restore_source_v3
      requestBody:
        $ref: '#/components/requestBodies/V3SourceRestoreRequest'
      responses:
        201:
          $ref: '#/components/responses/V3SourceData'
        400:
          $ref: '#/components/responses/BadRequest'
        409:
          $ref: '#/components/responses/ConflictingSourceName'
    delete:
      tags:
        - sources
        - v3
      summary: Delete a single source
      x-openapi-router-controller: cray.cfs.api.controllers.sources
      description:  >-
        Delete the given source. This will fail in any components are using
        the specified source.
      operationId: delete_source_v3
      responses:
        204:
          $ref: '#/components/responses/ResourceDeleted'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/ResourceNotFound'
    parameters:
      - name: source_id
        in: path
        description: Name of the target source
        required: true
        schema:
          type: string
          minLength: 1